refactor(rules): extract UntestedPublicAPIRule from testing_quality.go
testing_quality.go was 2109 lines covering ~19 testing-quality rules.
Step one of a god-file split: move UntestedPublicAPIRule and its 6
untestedPublicAPI* / testingQualityFilesByPath helpers into
testing_quality_untested_public_api.go (~140 LOC). The group is fully
self-contained — every helper is referenced only from this rule.

Brings the main file to 1983 LOC. Follow-ups can extract more rule
families; most share testingQuality* helpers and need separate split
planning.
